luna_study_ripam/requirements.txt
google-genai>=0.4.0
python-dotenv>=1.0.1
requests>=2.32.3
jsonschema>=4.23.0
rich>=13.8.1
